// fox-hunt-organization-front/src/utils/enums.ts
export enum organizationTypeEnum {
  FREE = 'FREE',
  PAID = 'PAID',
}

export enum organizationStatusEnum {
  NEW = 'NEW',
  DECLINED = 'DECLINED',
  ONBOARDING = 'ONBOARDING',
  ACTIVE = 'ACTIVE',
  DEACTIVATED = 'DEACTIVATED',
}

export enum userInvitationStatusEnum {
  NEW = 'NEW',
  ACCEPTED = 'ACCEPTED',
  DECLINED = 'DECLINED',
  EXPIRED = 'EXPIRED',
  FAILED = 'FAILED',
}

export enum UserFormEnum {
  firstName = 'firstName',
  lastName = 'lastName',
  dateOfBirth = 'dateOfBirth',
  city = 'city',
  email = 'email',
  country = 'country',
}

export enum organizationFormEnum {
  firstName = 'firstName',
  lastName = 'lastName',
  dateOfBirth = 'dateOfBirth',
  city = 'city',
  email = 'email',
  country = 'country',
}